Source: Created Source interface / refactor
Instead of having all these functions inside the Parser, we now abstract
out the lexemes into a Source, which you can then call the necessary
functions from to get the next Lexeme. This is much cleaner and makes us
able to handle whitespace without needing 4 helper functions in the
parser.