Parser: Fix error message being wrong for expressions
When we enter in "(1+" into the REPL (and hence, sent to the parser) the
first error message makes sense but the second error message use to say
"expecting operator". This patch implements a fix that removes this
behavior, as well as adding a test to ensure we do not regress backward.